fix(medios): una rafaga de fotos reescribe el catalogo una vez, no una por foto
Causa: cada foto recibida desde el movil dejaba el catalogo sucio y la
tarea de medios lo reescribia ENTERO (flexMsSave, escritura atomica)
1,5 s despues, entre foto y foto. Con 100 elementos son ~43 KB: una
decena de borrados de sector de 4 KB, cada uno con la cache apagada
decenas de ms -- tirones y, con el panel DSI sin CONFIG_LCD_DSI_ISR_
IRAM_SAFE, un destello cian por borrado (ver INSTALACION-USB-P4 5) --.
N fotos seguidas = N reescrituras.

Ahora la web marca la rafaga (progreso y publicacion de cada subida) y
el guardado espera a que acabe (3 s sin senales), con un tope duro:
lo sucio se guarda SIEMPRE antes de 20 s. Un corte de luz en ese
intervalo no pierde archivos: el recorrido del disco del arranque los
vuelve a registrar. Las descargas al movil no cuentan (no cambian el
catalogo).

Fallo encontrado por la prueba: las marcas de tiempo son millis() | 1
(0 = sin marca) y leidas en el MISMO milisegundo par van 1 ms por
delante; restadas sin signo daban 4.294.967.295 ms. La mitad de las
veces el guardado "agrupado" de antes se hacia en el acto. mlAge()
resta con signo y acota a 0. El mismo defecto en gedTicking() del
editor (un milisegundo creyendo que no estaba en pantalla) se corrige
igual.

Pruebas (test_ino):
- testGuardadoRafaga: la decision caso a caso, las senales salen de la
  web (y no de una descarga) y una rafaga de 5 fotos simulada con los
  tiempos reales: 1 reescritura (la politica anterior: 5).
- testVariasFotos: 8 fotos de 0,3 a 12 MP por el servidor web REAL
  (FlexOS_MediaWeb) con los callbacks de la placa y la Galeria abierta.
  La subida mas cara usa 525 KB de PSRAM (se lee por trozos: no crece
  con la foto); tras 8 fotos solo quedan +320 KB de miniaturas en cache
  (acotadas por ML_CACHE_N); ningun temporal; abrir las 8 en el visor y
  cerrarlo devuelve toda la PSRAM.

+++ FlexOS_Ultra/FlexOS_Ultra_GalleryEdit.h
@@ -1096,7 +1096,9 @@ static void gedCloseNow(){
 // ¿Se esta viendo? A pantalla completa lo dice el estado de la app; dentro
 // de una ventana de DeX, que su tick haya corrido hace nada.
 static uint32_t gedTickMs = 0;
-static bool gedTicking(){ return gedTickMs && millis() - gedTickMs < 300; }
+// Con signo: la marca es millis() | 1 y, leida en el mismo milisegundo par,
+// va 1 ms por delante (sin signo, "hace una eternidad": no se estaba viendo).
+static bool gedTicking(){ return gedTickMs && (int32_t)(millis() - gedTickMs) < 300; }
 static bool gedForeground(){
   return (gState == ST_APP && gAppId == IC_GALERIA && gAppState[IC_GALERIA] == ALIFE_RUNNING) || gedTicking();
 }
